50% audio coverage
