+++ audio/posFilter_test.go
@@ -1,0 +1,30 @@
+package audio
+
+import (
+	"testing"
+	"time"
+
+	"github.com/200sc/klangsynthese/synth"
+	"github.com/stretchr/testify/assert"
+)
+
+func TestPosFilter(t *testing.T) {
+	kla, err := synth.Int16.Sin()
+	assert.Nil(t, err)
+	x, y := new(float64), new(float64)
+	a := New(DefFont, kla.(Data), x, y)
+	x2 := 100.0
+	y2 := 100.0
+	DefFont.Filter(PosFilter(NewEars(&x2, &y2, 100, 300)))
+	err = <-a.Play()
+	time.Sleep(a.PlayLength())
+	// Assert left ear hears audio
+	x2 -= 200
+	err = <-a.Play()
+	time.Sleep(a.PlayLength())
+	// Assert right ear hears audio
+	y2 += 500
+	err = <-a.Play()
+	time.Sleep(a.PlayLength())
+	// Assert nothing is heard
+}
